# social-media/backend/docs/swagger-auth/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/internal_handlers.UserResponse'
    type: object
...
  internal_handlers.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  internal_handlers.RegisterRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
  internal_handlers.ResendVerificationRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_handlers.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  internal_handlers.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      display_name:
        type: string
    type: object
...
  internal_handlers.UserResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Authentication service for SocialHub platform
  title: SocialHub Auth API
  version: 1.0.0
paths:
...
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset email
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Forgot password
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      description: Invalidate user session
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /api/v1/auth/me:
    get:
      description: Get current authenticated user details
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.UserResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /api/v1/auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend verification email to user
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Resend verification email
      tags:
      - auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password with token from email
      parameters:
      - description: Token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password
      tags:
      - auth
  /api/v1/auth/verify/{token}:
    get:
      description: Verify user email with token
      parameters:
      - description: Verification token
        in: path
        name: token
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify email
      tags:
      - auth
...
  /api/v1/user/profile:
    put:
      consumes:
      - application/json
      description: Update current user profile
      parameters:
      - description: Profile update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.UserResponse'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - auth
...
swagger: "2.0"